tech: Move matchers to PactSwift
