+++ Sources/Model/HeaderBuilder.swift
@@ -1,0 +1,39 @@
+//
+//  Created by Oliver Jones on 9/1/2023.
+//  Copyright © 2023 Oliver Jones. All rights reserved.
+//
+//  Permission to use, copy, modify, and/or distribute this software for any
+//  purpose with or without fee is hereby granted, provided that the above
+//  copyright notice and this permission notice appear in all copies.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+//  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+//  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
+//  SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+//  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+//  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
+//  IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+//
+
+import Foundation
+
+public protocol HeaderBuilder {
+           
+    /// Configures a header for the Interaction.
+    ///
+    /// To include matching rules for the header, include the matching rule JSON format with the value as a single JSON document. I.e.
+    ///
+    /// ```
+    /// let value = #"{"value":"2", "pact:matcher:type":"regex", "regex":"\\d+"}"#
+    /// builder.header("id", [value]);
+    /// ```
+    /// See [IntegrationJson.md](https://github.com/pact-foundation/pact-reference/blob/master/rust/pact_ffi/IntegrationJson.md)
+    ///
+    /// - Throws: ``Interaction/Error/canNotBeModified`` if the interaction or Pact can't be modified (i.e. the mock server for it has already started).
+    /// - Parameters:
+    ///   - name: The header name.
+    ///   - value: The header values.
+    ///
+    @discardableResult
+    func header(_ name: String, values: [String]) throws -> Self
+}
