--- Sources/Model/Matchers/GeneratorType.swift
@@ -1,33 +1,0 @@
-//
-//  Created by Oliver Jones on 16/12/2022.
-//  Copyright © 2022 Oliver Jones. All rights reserved.
-//
-//  Permission to use, copy, modify, and/or distribute this software for any
-//  purpose with or without fee is hereby granted, provided that the above
-//  copyright notice and this permission notice appear in all copies.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-//  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
-//  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
-//  SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-//  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
-//  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
-//  IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-//
-
-import Foundation
-
-enum GeneratorType: String, Encodable {
-    case randomInt = "RandomInt"
-    case uuid = "Uuid"
-    case randomDecimal = "RandomDecimal"
-    case randomHex = "RandomHexadecimal"
-    case randomString = "RandomString"
-    case regex = "Regex"
-    case date = "Date"
-    case time = "Time"
-    case dateTime = "DateTime"
-    case randomBoolean = "RandomBoolean"
-    case providerState = "ProviderState"
-    case mockServerUrl = "MockServerURL"
-}
