--- Sources/Model/Matchers/RegexMatcher.swift
@@ -1,44 +1,0 @@
-//
-//  Created by Oliver Jones on 16/12/2022.
-//  Copyright © 2022 Oliver Jones. All rights reserved.
-//
-//  Permission to use, copy, modify, and/or distribute this software for any
-//  purpose with or without fee is hereby granted, provided that the above
-//  copyright notice and this permission notice appear in all copies.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-//  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
-//  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
-//  SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-//  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
-//  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
-//  IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-//
-
-import Foundation
-
-public struct RegexMatcher: Matcher {
-    
-    let type = "regex"
-    let value: String
-    let regex: String
-    
-    init(example: String, regex: String) {
-        self.value = example
-        self.regex = regex
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case type = "pact:matcher:type"
-        case value
-        case regex
-    }
-}
-
-extension Matcher where Self == RegexMatcher {
-    
-    public static func regex(_ regex: String, example: String) -> Self {
-        Self(example: example, regex: regex)
-    }
-
-}
